consolidate settings actions
